Minor fixes & dismembering mapper class
