+++ megalista_dataflow/mappers/ads_user_list_pii_hashing_mapper.py
@@ -15,38 +15,14 @@
 import logging
 
 from models.execution import Batch
+from mappers.abstract_list_pii_hashing_mapper import ListPIIHashingMapper
 
-
-class FieldHasher:
-    def __init__(self, should_hash_fields):
-        self.should_hash_fields = should_hash_fields
-
-    def hash_field(self, field):
-        import hashlib
-
-        if self.should_hash_fields:
-            return hashlib.sha256(field.strip().lower().encode("utf-8")).hexdigest()
-
-        return field
-
-
-def _is_data_present(dict, key):
-    return key in dict and dict[key] is not None and dict[key] != ""
-
-
-class AdsUserListPIIHashingMapper:
+class AdsUserListPIIHashingMapper(ListPIIHashingMapper):
     def __init__(self):
         self.logger = logging.getLogger("megalista.AdsUserListPIIHashingMapper")
 
     def _hash_user(self, user, hasher):
-        hashable_keys = ("email",
-                         "mailing_address_first_name",
-                         "mailing_address_last_name",
-                         "mailing_address_country",
-                         "mailing_address_zip",
-                         "phone",
-                         "mobile_device_id",
-                         "user_id")
+        hashable_keys = self._get_default_hasheable_keys()
         processed_user = {}
         # include non PII keys as is (these should not be hashed)
         for k, v in user.items():
@@ -54,17 +30,17 @@ class AdsUserListPIIHashingMapper:
             processed_user[k] = v
 
         try:
-            if _is_data_present(user, "email"):
+            if self._is_data_present(user, "email"):
                 processed_user["hashed_email"] = hasher.hash_field(user["email"])
         except:
             self.logger.error(f"Error hashing email for user: {str(user)}")
 
         try:
             if (
-                _is_data_present(user, "mailing_address_first_name")
-                and _is_data_present(user, "mailing_address_last_name")
-                and _is_data_present(user, "mailing_address_country")
-                and _is_data_present(user, "mailing_address_zip")
+                self._is_data_present(user, "mailing_address_first_name")
+                and self._is_data_present(user, "mailing_address_last_name")
+                and self._is_data_present(user, "mailing_address_country")
+                and self._is_data_present(user, "mailing_address_zip")
             ):
                 processed_user["address_info"] = {
                     "hashed_first_name": hasher.hash_field(
@@ -80,46 +56,18 @@ class AdsUserListPIIHashingMapper:
             self.logger.error(f"Error hashing address for user: {str(user)}")
 
         try:
-            if _is_data_present(user, "phone"):
+            if self._is_data_present(user, "phone"):
                 processed_user["hashed_phone_number"] = hasher.hash_field(user["phone"])
         except:
             self.logger.error(f"Error hashing phone for user: {str(user)}")
 
-        if _is_data_present(user, "mobile_device_id"):
+        if self._is_data_present(user, "mobile_device_id"):
             processed_user["mobile_id"] = user["mobile_device_id"]
 
         try:
-            if _is_data_present(user, "user_id"):
+            if self._is_data_present(user, "user_id"):
                 processed_user["third_party_user_id"] = hasher.hash_field(user["user_id"])
         except:
             self.logger.error(f"Error hashing user_id for user: {str(user)}")
 
         return processed_user
-
-    def _get_should_hash_fields(self, metadata_list):
-
-        if len(metadata_list) < 3:
-            return True
-
-        should_hash_fields = metadata_list[2]
-
-        if not should_hash_fields:
-            return True
-
-        return should_hash_fields.lower() != "false"
-
-    def hash_users(self, batch: Batch):
-
-        should_hash_fields = self._get_should_hash_fields(
-            batch.execution.destination.destination_metadata
-        )
-        self.logger.debug(f"Should hash fields is {str(should_hash_fields)}")
-
-        hashed_elements = [
-            self._hash_user(element, FieldHasher(should_hash_fields))
-            for element in batch.elements
-        ]
-
-        return Batch(
-            batch.execution, [element for element in hashed_elements if element]
-        )
