+++ megalista_dataflow/mappers/dv_user_list_pii_hashing_mapper.py
@@ -15,38 +15,15 @@
 import logging
 
 from models.execution import Batch
+from mappers.abstract_list_pii_hashing_mapper import ListPIIHashingMapper
 
 
-class FieldHasher:
-    def __init__(self, should_hash_fields):
-        self.should_hash_fields = should_hash_fields
-
-    def hash_field(self, field):
-        import hashlib
-
-        if self.should_hash_fields:
-            return hashlib.sha256(field.strip().lower().encode("utf-8")).hexdigest()
-
-        return field
-
-
-def _is_data_present(dict, key):
-    return key in dict and dict[key] is not None and dict[key] != ""
-
-
-class DVUserListPIIHashingMapper:
+class DVUserListPIIHashingMapper(ListPIIHashingMapper):
     def __init__(self):
         self.logger = logging.getLogger("megalista.DVUserListPIIHashingMapper")
 
     def _hash_user(self, user, hasher):
-        hashable_keys = ("email",
-                         "mailing_address_first_name",
-                         "mailing_address_last_name",
-                         "mailing_address_country",
-                         "mailing_address_zip",
-                         "phone",
-                         "mobile_device_id",
-                        )
+        hashable_keys = self._get_default_hasheable_keys()
         processed_user = {}
         # include non PII keys as is (these should not be hashed)
         for k, v in user.items():
@@ -54,17 +31,17 @@ class DVUserListPIIHashingMapper:
             processed_user[k] = v
 
         try:
-            if _is_data_present(user, "email"):
+            if self._is_data_present(user, "email"):
                 processed_user["hashedEmails"] = hasher.hash_field(user["email"])
         except:
             self.logger.error(f"Error hashing email for user: {str(user)}")
 
         try:
             if (
-                _is_data_present(user, "mailing_address_first_name")
-                and _is_data_present(user, "mailing_address_last_name")
-                and _is_data_present(user, "mailing_address_country")
-                and _is_data_present(user, "mailing_address_zip")
+                self._is_data_present(user, "mailing_address_first_name")
+                and self._is_data_present(user, "mailing_address_last_name")
+                and self._is_data_present(user, "mailing_address_country")
+                and self._is_data_present(user, "mailing_address_zip")
             ):
                 processed_user["hashedFirstName"] = hasher.hash_field(user["mailing_address_first_name"])
                 processed_user["hashedLastName"] = hasher.hash_field(user["mailing_address_last_name"])
@@ -75,40 +52,12 @@ class DVUserListPIIHashingMapper:
             self.logger.error(f"Error hashing address for user: {str(user)}")
 
         try:
-            if _is_data_present(user, "phone"):
+            if self._is_data_present(user, "phone"):
                 processed_user["hashedPhoneNumbers"] = hasher.hash_field(user["phone"])
         except:
             self.logger.error(f"Error hashing phone for user: {str(user)}")
 
-        if _is_data_present(user, "mobile_device_id"):
+        if self._is_data_present(user, "mobile_device_id"):
             processed_user["mobileDeviceIds"] = user["mobile_device_id"]
 
         return processed_user
-
-    def _get_should_hash_fields(self, metadata_list):
-
-        if len(metadata_list) < 3:
-            return True
-
-        should_hash_fields = metadata_list[2]
-
-        if not should_hash_fields:
-            return True
-
-        return should_hash_fields.lower() != "false"
-
-    def hash_users(self, batch: Batch):
-
-        should_hash_fields = self._get_should_hash_fields(
-            batch.execution.destination.destination_metadata
-        )
-        self.logger.debug(f"Should hash fields is {str(should_hash_fields)}")
-
-        hashed_elements = [
-            self._hash_user(element, FieldHasher(should_hash_fields))
-            for element in batch.elements
-        ]
-
-        return Batch(
-            batch.execution, [element for element in hashed_elements if element]
-        )
